Add api.s type and revert JSX's d.ts rename
