+++ packages/sinuous/hydrate/src/index.d.ts
@@ -1,4 +1,4 @@
-import { JSXInternal } from '../../jsx-internal';
+import { JSXInternal } from '../../jsx';
 import { ElementChildren, FunctionComponent } from '../../shared';
 
 interface VNode<P = {}> {
