+++ packages/sinuous/src/index.d.ts
@@ -1,7 +1,7 @@
 export = sinuous;
 export as namespace sinuous;
 
-import { JSXInternal } from '../jsx-internal';
+import { JSXInternal } from '../jsx';
 import { HyperscriptApi } from '../h/src';
 import * as _shared from '../shared'
 import * as _o from '../observable/src';
